requirements/requirements.txt
Django==1.6.2
Fabric==1.8.2
MySQL-python==1.2.5
Pillow==2.3.0
PyYAML==3.10
South==0.8.4
Unipath==1.0
amqp==1.4.3
anyjson==0.3.3
argh==0.24.1
billiard==3.3.0.16
celery==3.1.9
django-celery==3.1.9
django-ckeditor-updated==4.2.6
django-filebrowser==3.5.3
django-grappelli==2.5.1
django-reversion==1.8.0
django-supervisor==0.3.2
ecdsa==0.10
gunicorn==18.0
httplib2==0.8
kombu==3.0.12
meld3==0.6.10
oauth2==1.5.211
paramiko==1.12.2
pathtools==0.1.2
pycrypto==2.6.1
python-instagram==0.8.0
pytz==2013.9
simplejson==3.3.2
sqlparse==0.1.11
supervisor==3.0
watchdog==0.7.1
wsgiref==0.1.2
